+++ basicalter/slice_test.go
@@ -8,6 +8,42 @@ import (
 	"github.com/jeremmfr/go-utils/basiccheck"
 )
 
+func TestUniqueInSlice(t *testing.T) {
+	sliceOfString := []string{"foo", "bar"}
+	if len(basicalter.UniqueInSlice(sliceOfString)) != len(sliceOfString) {
+		t.Errorf("UniqueInSlice returned bad slice with slice without duplicate entry")
+	}
+
+	sliceOfString = append(sliceOfString, "foo")
+	if v := basicalter.UniqueInSlice(sliceOfString); len(v) == len(sliceOfString) {
+		t.Errorf("UniqueInSlice didn't remove duplicate foo in slice: %v", v)
+	}
+
+	// test empty slice
+	_ = len(basicalter.UniqueInSlice([]string{}))
+
+	// test nil slice
+	var nilSlice []string
+	if basicalter.UniqueInSlice(nilSlice) != nil {
+		t.Errorf("UniqueInSlice didn't return nil slice with nil input slice")
+	}
+
+	sliceOfInt := []int{1, 2}
+	if len(basicalter.UniqueInSlice(sliceOfInt)) != len(sliceOfInt) {
+		t.Errorf("UniqueInSlice returned bad slice with slice without duplicate entry")
+	}
+
+	sliceOfInt = append(sliceOfInt, 1)
+	if v := basicalter.UniqueInSlice(sliceOfInt); len(v) == len(sliceOfInt) {
+		t.Errorf("UniqueInSlice didn't remove duplicate 1 in slice: %v", v)
+	}
+
+	// test slice with one element
+	if v := basicalter.UniqueInSlice([]float64{1.2}); len(v) != 1 || v[0] != 1.2 {
+		t.Errorf("UniqueInSlice didn't return the same slice with one element")
+	}
+}
+
 func TestUniqueStrings(t *testing.T) {
 	sliceOfString := []string{"foo", "bar"}